Update Codegen scripts to not try to import types from missing pkg
These changes were necessary for me to run codegen script in dev repo.  I wasn't sure whether to submit this to major, minor, master, or all.

The @graphql-codegen/plugin-helpers package that was referenced is not installed in the project.  It seems the more recent way to get the config object type is to import it from the cli package.